Renamed and fixed CQLLibraryDelete.cy.ts and updated DeleteCQLLibrary.cy.ts
